+++ services/galley/src/Galley/App.hs
@@ -67,6 +67,7 @@ import qualified Galley.Aws               as Aws
 import qualified Galley.Queue             as Q
 import qualified OpenSSL.X509.SystemStore as Ssl
 import qualified System.Logger            as Logger
+import qualified System.Logger.Extended   as Logger
 
 data DeleteItem = TeamItem TeamId UserId (Maybe ConnId)
     deriving (Eq, Ord, Show)
@@ -123,16 +124,9 @@ instance MonadHttp Galley where
 instance HasRequestId Galley where
     getRequestId = view reqId
 
-mkLogger :: Opts -> IO Logger
-mkLogger opts = Logger.new $ Logger.defSettings
-    & Logger.setLogLevel (opts ^. optLogLevel)
-    & Logger.setOutput Logger.StdOut
-    & Logger.setFormat Nothing
-    & Logger.setNetStrings (opts ^. optLogNetStrings)
-
 createEnv :: Metrics -> Opts -> IO Env
 createEnv m o = do
-    l   <- mkLogger o
+    l   <- Logger.mkLogger (o ^. optLogLevel) (o ^. optLogNetStrings)
     mgr <- initHttpManager o
     Env def m o l mgr <$> initCassandra o l
                       <*> Q.new 16000
@@ -144,8 +138,9 @@ initCassandra o l = do
     c <- maybe (C.initialContactsPlain (o^.optCassandra.casEndpoint.epHost))
                (C.initialContactsDisco "cassandra_galley")
                (unpack <$> o^.optDiscoUrl)
-    C.init (Logger.clone (Just "cassandra.galley") l) $
-              C.setContacts (NE.head c) (NE.tail c)
+    C.init
+            . C.setLogger (C.mkLogger (Logger.clone (Just "cassandra.galley") l))
+            . C.setContacts (NE.head c) (NE.tail c)
             . C.setPortNumber (fromIntegral $ o^.optCassandra.casEndpoint.epPort)
             . C.setKeyspace (Keyspace $ o^.optCassandra.casKeyspace)
             . C.setMaxConnections 4