+++ services/spar/src/Spar/Scim/User.hs
@@ -25,7 +25,6 @@ import Imports
 import Brig.Types.User       as Brig
 import Control.Lens hiding ((.=), Strict)
 import Control.Monad.Except
-import Control.Monad.Extra (whenM)
 import Crypto.Hash
 import Data.Aeson as Aeson
 import Data.Id
@@ -33,7 +32,8 @@ import Data.Range
 import Data.String.Conversions
 import Galley.Types.Teams    as Galley
 import Network.URI
-import Spar.App (Spar, Env, wrapMonadClient, sparCtxOpts, createUser_, wrapMonadClient)
+
+import Spar.App (Spar, Env, wrapMonadClient, sparCtxOpts, sparCtxLogger, createUser_, wrapMonadClient)
 import Spar.Intra.Galley
 import Spar.Scim.Types
 import Spar.Scim.Auth ()
@@ -45,6 +45,7 @@ import qualified SAML2.WebSSO as SAML
 import qualified Spar.Data    as Data
 import qualified Spar.Intra.Brig as Intra.Brig
 import qualified URI.ByteString as URIBS
+import qualified System.Logger as Log
 
 import qualified Web.Scim.Class.User              as Scim
 import qualified Web.Scim.Filter                  as Scim
@@ -71,7 +72,7 @@ instance Scim.UserDB Spar where
     members <- lift $ getTeamMembers stiTeam
     brigusers :: [User]
       <- filter (not . userDeleted) <$>
-         lift (Intra.Brig.getUsers ((^. Galley.userId) <$> members))
+         lift (Intra.Brig.getBrigUsers ((^. Galley.userId) <$> members))
     scimusers :: [Scim.StoredUser ScimUserExtra]
       <- lift . wrapMonadClient . Data.getScimUsers $ Brig.userId <$> brigusers
     let check user = case mbFilter of
@@ -91,7 +92,7 @@ instance Scim.UserDB Spar where
       -> Scim.ScimHandler Spar (Maybe (Scim.StoredUser ScimUserExtra))
   get ScimTokenInfo{stiTeam} uidText = do
     uid <- parseUid uidText
-    mbBrigUser <- lift (Intra.Brig.getUser uid)
+    mbBrigUser <- lift (Intra.Brig.getBrigUser uid)
     if isJust mbBrigUser && (userTeam =<< mbBrigUser) == Just stiTeam
       then lift . wrapMonadClient . Data.getScimUser $ uid
       else pure Nothing
@@ -111,12 +112,35 @@ instance Scim.UserDB Spar where
     updateValidScimUser tokinfo uidText =<< validateScimUser tokinfo newScimUser
 
   delete :: ScimTokenInfo -> Text -> Scim.ScimHandler Spar Bool
-  delete _ _ =
-      throwError $ Scim.ScimError
-          mempty
-          (Scim.Status 404)
-          Nothing
-          (Just "User delete is not implemented yet")  -- TODO
+  delete ScimTokenInfo{stiTeam} uidText = do
+    uid :: UserId <- parseUid uidText
+    mbBrigUser <- lift (Intra.Brig.getBrigUser uid)
+    case mbBrigUser of
+      Nothing -> do
+        -- double-deletion gets you a 404.
+        throwError $ Scim.notFound "user" (cs $ show uid)
+      Just brigUser -> do
+        -- FUTUREWORK: currently it's impossible to delete the last available team owner via SCIM
+        -- (because that owner won't be managed by SCIM in the first place), but if it ever becomes
+        -- possible, we should do a check here and prohibit it.
+        unless (userTeam brigUser == Just stiTeam) $
+        -- users from other teams get you a 404.
+          throwError $ Scim.notFound "user" (cs $ show uid)
+        ssoId <- maybe (logThenServerError $ "no userSSOId for user " <> cs uidText)
+                       pure
+                       $ Brig.userSSOId brigUser
+        uref <- either logThenServerError pure $ Intra.Brig.fromUserSSOId ssoId
+        lift . wrapMonadClient $ Data.deleteSAMLUser uref
+        lift . wrapMonadClient $ Data.deleteScimUser uid
+        lift $ Intra.Brig.deleteBrigUser uid
+        return True
+          where
+            logThenServerError :: String -> Scim.ScimHandler Spar b
+            logThenServerError err = do
+              logger <- asks sparCtxLogger
+              Log.err logger $ Log.msg err
+              throwError $ Scim.serverError "Server Error"
+
 
   getMeta :: ScimTokenInfo -> Scim.ScimHandler Spar Scim.Meta
   getMeta _ =
@@ -187,6 +211,9 @@ validateScimUser' idp richInfoLimit user = do
     handl <- validateHandle (Scim.userName user)
     mbName <- mapM validateName (Scim.displayName user)
     richInfo <- validateRichInfo (Scim.extra user ^. sueRichInfo)
+
+    -- NB: We assume that checking that the user does _not_ exist has already been done before;
+    -- the hscim library check does a 'get' before a 'create'.
     pure $ ValidScimUser user uref handl mbName richInfo
 
   where
@@ -241,15 +268,15 @@ createValidScimUser
   :: forall m. (m ~ Scim.ScimHandler Spar)
   => ValidScimUser -> m (Scim.StoredUser ScimUserExtra)
 createValidScimUser (ValidScimUser user uref handl mbName richInfo) = do
+
     -- FUTUREWORK: The @hscim@ library checks that the handle is not taken before 'create' is
     -- even called. However, it does that in an inefficient manner. We should remove the check
     -- from @hscim@ and do it here instead.
 
-    -- Check that the UserRef is not taken.
-    whenM (isJust <$> lift (wrapMonadClient (Data.getUser uref))) $
-        throwError Scim.conflict {Scim.detail = Just "externalId is already taken"}
     -- Generate a UserId will be used both for scim user in spar and for brig.
     buid <- Id <$> liftIO UUID.nextRandom
+    assertUserRefUnused buid uref
+
     -- Create SCIM user here in spar.
     storedUser <- lift $ toScimStoredUser buid user
     lift . wrapMonadClient $ Data.insertScimUser buid storedUser
@@ -262,9 +289,9 @@ createValidScimUser (ValidScimUser user uref handl mbName richInfo) = do
     lift $ createUser_ buid uref mbName ManagedByScim
     -- Set user handle on brig (which can't be done during user creation yet).
     -- TODO: handle errors better here?
-    lift $ Intra.Brig.setHandle buid handl
+    lift $ Intra.Brig.setBrigUserHandle buid handl
     -- Set rich info on brig
-    lift $ Intra.Brig.setRichInfo buid richInfo
+    lift $ Intra.Brig.setBrigUserRichInfo buid richInfo
 
     pure storedUser
 
@@ -293,6 +320,9 @@ updateValidScimUser tokinfo uidText newScimUser = do
       <- let err = throwError $ Scim.notFound "user" uidText
          in maybe err pure =<< Scim.get tokinfo uidText
 
+    let userRef = newScimUser ^. vsuSAMLUserRef
+    assertUserRefUnused uid userRef
+
     if Scim.value (Scim.thing oldScimStoredUser) == (newScimUser ^. vsuUser)
       then pure oldScimStoredUser
       else do
@@ -301,16 +331,16 @@ updateValidScimUser tokinfo uidText newScimUser = do
 
         -- update 'SAML.UserRef'
         let uref = newScimUser ^. vsuSAMLUserRef
-        lift . wrapMonadClient $ Data.insertUser uref uid  -- on spar
-        bindok <- lift $ Intra.Brig.bindUser uid uref  -- on brig
+        lift . wrapMonadClient $ Data.insertSAMLUser uref uid  -- on spar
+        bindok <- lift $ Intra.Brig.bindBrigUser uid uref  -- on brig
         unless bindok . throwError $
             Scim.serverError "Failed to update SAML UserRef on brig."
             -- this can only happen if user is found in spar.scim_user, but missing on brig.
             -- (internal error?  race condition?)
 
-        maybe (pure ()) (lift . Intra.Brig.setName uid) $ newScimUser ^. vsuName
-        lift . Intra.Brig.setHandle uid $ newScimUser ^. vsuHandle
-        lift . Intra.Brig.setRichInfo uid $ newScimUser ^. vsuRichInfo
+        maybe (pure ()) (lift . Intra.Brig.setBrigUserName uid) $ newScimUser ^. vsuName
+        lift . Intra.Brig.setBrigUserHandle uid $ newScimUser ^. vsuHandle
+        lift . Intra.Brig.setBrigUserRichInfo uid $ newScimUser ^. vsuRichInfo
 
         -- store new user value to scim_user table (spar). (this must happen last, so in case
         -- of crash the client can repeat the operation and it won't be considered a noop.)
@@ -402,6 +432,23 @@ calculateVersion uidText usr = Scim.Weak (Text.pack (show h))
   where
     h :: Digest SHA256
     h = hashlazy (Aeson.encode (Scim.WithId uidText usr))
+
+{-|
+Check that the UserRef is not taken; or that it's taken by the given user id.
+
+ASSUMPTION: every scim user has a 'SAML.UserRef', and the `SAML.NameID` in it corresponds
+to a single `externalId`.
+-}
+assertUserRefUnused :: UserId -> SAML.UserRef -> Scim.ScimHandler Spar ()
+assertUserRefUnused wireUserId userRef = do
+  mExistingUserId <- lift $ wrapMonadClient (Data.getSAMLUser userRef)
+  case mExistingUserId of
+    -- No existing user for this userRef; it's okay to set it
+    Nothing -> return ()
+    -- A user exists; verify that it's the same user before updating
+    Just existingUserId  ->
+      unless (existingUserId == wireUserId) $
+        throwError Scim.conflict {Scim.detail = Just "externalId is already taken"}
 
 {- TODO: might be useful later.
 ~~~~~~~~~~~~~~~~~~~~~~~~~